#69 prevent removal of TAs when not logged into the system

+++ backend/submit/remove-ta.php
@@ -2,6 +2,8 @@
 $course = $_POST['course'];
 $id = $_POST["id"];
 include $_SERVER['DOCUMENT_ROOT'] . '/CSE442-542/2018-Summer/team03/backend/session/session.php';
-
+if (!$logged_in) {
+  exit();
+}
 $database->removeTA($id, $course);
 exit();
